[ALPS08461551] ISP8 P2 QOF Porting and traw qof on
brief: ISP8 P2 Qof porting and traw qof on

MTK-Commit-Id: 8ebd545bf270c8be493e78aab20f6930f8d8459f

CR-Id: ALPS08461551
Feature: VMM

+++ imgsys/imgsys/isp8/mtk_imgsys-dev.h
@@ -406,7 +406,7 @@ struct mtk_imgsys_dev {
 	int (*is_singledev_mode)(struct mtk_imgsys_request *req);
 #endif
 	/* IMGSYS Qucik on off */
-	u8 qof_ver;
+	u32 qof_ver;
 	u32 *work_buf_va;
 	dma_addr_t work_buf_pa;
 	u32 *traw_work_buf_va;
